pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "specsy"
version = "0.6.4"
readme = "README.rst"
requires-python = ">=3.11"
license = "GPL-3.0-or-later"
authors = [{name = "Vital Fernández", email = "vgf@umich.edu"}]
description = "Model fitting package for the chemical analysis of astronomical spectra"
dependencies = ["arviz~=0.21",
                "corner~=2.2",
                "h5netcdf~=1.3.0",
                "jax~=0.4",
                "jaxlib~=0.4",
                "lime-stable~=2.0.dev6",
                "lmfit~=1.3",
                "bokeh~=3.7",
                "matplotlib~=3.10",
                "numpy~=2.2",
                "pymc~=5.1",
                "scipy~=1.15",
                "toml~=0.10",
                "tomli >= 2.0.0 ; python_version < '3.11'",
                "xarray~=2024.3.0"]

classifiers = ["Programming Language :: Python :: 3",
               "Programming Language :: Python :: 3.11"]

[project.optional-dependencies]
tensors = ["pytensor~=2.20"]

docs = ["nbsphinx~=0.9",
        "ipympl~=0.9",
        "sphinx-rtd-theme~=3.0"]

tests = ["pytest~=8.3",
         "pytest-cov~=6.0",
         "pytest-mpl~=0.17"]

[tool.pytest.ini_options]
pythonpath = ["src"]
mpl-baseline-path = 'tests/baseline'
mpl-results-path = 'tests/outputs'
mpl-results-always = false
addopts = "-p no:asdf_schema_tester"
